sonar - squid:S2164 - Math should not be performed on floats
git-svn-id: http://josm.openstreetmap.de/svn/trunk@10303 0c6e7542-c601-0410-84e7-c038aed88b3b

+++ src/org/openstreetmap/josm/data/osm/visitor/paint/StyledMapRenderer.java
@@ -1502,7 +1502,7 @@ public class StyledMapRenderer extends AbstractMapRenderer {
 
                             while (dist < segmentLength) {
                                 for (int i = 0; i < 2; ++i) {
-                                    float onewaySize = i == 0 ? 3f : 2f;
+                                    double onewaySize = i == 0 ? 3d : 2d;
                                     GeneralPath onewayPath = i == 0 ? onewayArrowsCasing : onewayArrows;
 
                                     // scale such that border is 1 px
